Update movie item view data model to item view data model

+++ lib/data/source/network/response/tv/tv_item_response.dart
@@ -24,7 +24,7 @@ class TvItemResponse {
   @JSONField(name: "poster_path")
   late String posterPath;
   @JSONField(name: "vote_average")
-  late int voteAverage;
+  late double voteAverage;
   @JSONField(name: "vote_count")
   late int voteCount;
